Experimental switch to Ubuntu 16.10 and Mapnik 3
Mapnik 3 support for Ocitysmap rendering was added by Étienne Loks

http://lists.nongnu.org/archive/html/maposmatic-dev/2016-10/msg00000.html

The move to a more recent Ubuntu version was needed as prior to 16.10
mapnik and python-mapnik system packages were not built with Cairo
suppprt, which is required by Ocitysmap

Unfortunately text lables are currently screwed up in CartoOSM style,
I don't really now why a wrong font size seems to be picked yet.
Funny enough this is only a problem with CartoOSM and not with
the current german style which is derived from it.

Maybe it is a problem in th latest CartoOSM git HEAD after all?