grow partition first, then the file system within
vagrant-disksize sometimes seems to increase the vdi file size only

So we manually grow the partition, and only then the file system,
as part of provisioning now. In the end this is going to be an
indempotent operation if vagrant-disksize actually did the job already;
making this a belt-and-suspenders approach.